+++ src/libs/satellite/src/usage/assert.rs
@@ -1,4 +1,6 @@
-use crate::usage::user_usage::{increase_db_usage, is_db_collection_no_usage};
+use crate::types::state::CollectionType;
+use crate::usage::state::update_user_usage;
+use crate::usage::utils::is_db_collection_no_usage;
 use junobuild_collections::types::core::CollectionKey;
 use junobuild_shared::controllers::is_controller;
 use junobuild_shared::types::state::{Controllers, UserId};
@@ -18,7 +20,7 @@ pub fn increment_and_assert_db_usage(
         return Ok(());
     }
 
-    let user_usage = increase_db_usage(collection, &caller);
+    let user_usage = update_user_usage(collection, &CollectionType::Db, &caller, None);
 
     if let Some(max_changes_per_user) = max_changes_per_user {
         if user_usage.changes_count > max_changes_per_user {
